requirements.txt
django>=4.2
psycopg2-binary
python-telegram-bot>=13.0
django-jsoneditor
djangorestframework
